requirements.txt
roboflow
ultralytics
python-dotenv
